refactor: ♻️ explode ui library into main application
🏁 Closes: #489
