+++ apps/web/src/components/table/table.tsx
@@ -4,7 +4,6 @@ import type {
   SortingState,
 } from "@tanstack/react-table";
 
-import { cn } from "@popcorn.fyi/ui/utils";
 import {
   flexRender,
   getCoreRowModel,
@@ -16,6 +15,7 @@ import {
 } from "@tanstack/react-table";
 import { useState } from "react";
 
+import { cn } from "@/lib/cn";
 import { fuzzyFilter } from "@/lib/fuzzy-filter";
 
 import { TableGlobalFilter } from "./table-global-filter";
