--- libs/ui/README.md
@@ -1,1 +1,0 @@
-# @popcorn.fyi/ui
