--- libs/ui/eslint.config.ts
@@ -1,1 +1,0 @@
-export { default } from "@popcorn.fyi/eslint";
