--- libs/ui/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@popcorn.fyi/ui",
-  "private": true,
-  "exports": {
-    "./utils": "./src/utils.ts"
-  },
-  "scripts": {
-    "build": "storybook build",
-    "coverage": "vitest run --coverage",
-    "dev": "storybook dev -p 6006 --no-open",
-    "lint": "eslint --cache .",
-    "test": "vitest",
-    "typecheck": "tsc"
-  },
-  "dependencies": {
-    "clsx": "2.1.1",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "tailwind-merge": "3.3.1"
-  },
-  "devDependencies": {
-    "@popcorn.fyi/eslint": "workspace:*",
-    "@popcorn.fyi/tailwind": "workspace:*",
-    "@popcorn.fyi/ts": "workspace:*",
-    "@storybook/addon-a11y": "9.0.14",
-    "@storybook/addon-docs": "9.0.14",
-    "@storybook/addon-themes": "9.0.14",
-    "@storybook/react-vite": "9.0.14",
-    "@tailwindcss/postcss": "4.1.11",
-    "@testing-library/jest-dom": "6.6.3",
-    "@types/react": "19.1.8",
-    "@types/react-dom": "19.1.6",
-    "happy-dom": "18.0.1",
-    "postcss": "8.5.6",
-    "storybook": "9.0.14"
-  }
-}
