--- libs/ui/turbo.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "extends": ["//"],
-  "tasks": {
-    "build": {
-      "outputs": ["storybook-static/**"]
-    }
-  }
-}
